requirements.txt
dearpygui>=1.10.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
networkx>=3.2
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
